// src/type.ts
    export interface Name {
        common: string;
        official: string;
        nativeName: NativeName;
    }

    export interface EUR {
        name: string;
        symbol: string;
    }

    export interface NZD {
        name: string;
        symbol: string;
    }

    export interface CUC {
        name: string;
        symbol: string;
    }

    export interface CUP {
        name: string;
        symbol: string;
    }

    export interface VES {
        name: string;
        symbol: string;
    }

    export interface USD {
        name: string;
        symbol: string;
    }

    export interface WST {
        name: string;
        symbol: string;
    }

    export interface KHR {
        name: string;
        symbol: string;
    }

    export interface HKD {
        name: string;
        symbol: string;
    }

    export interface MRU {
        name: string;
        symbol: string;
    }

    export interface YER {
        name: string;
        symbol: string;
    }

    export interface DJF {
        name: string;
        symbol: string;
    }

    export interface EGP {
        name: string;
        symbol: string;
    }

    export interface HRK {
        name: string;
        symbol: string;
    }

    export interface CHF {
        name: string;
        symbol: string;
    }

    export interface KZT {
        name: string;
        symbol: string;
    }

    export interface DKK {
        name: string;
        symbol: string;
    }

    export interface XOF {
        name: string;
        symbol: string;
    }

    export interface BMD {
        name: string;
        symbol: string;
    }

    export interface SHP {
        name: string;
        symbol: string;
    }

    export interface ZAR {
        name: string;
        symbol: string;
    }

    export interface RWF {
        name: string;
        symbol: string;
    }

    export interface MOP {
        name: string;
        symbol: string;
    }

    export interface BIF {
        name: string;
        symbol: string;
    }

    export interface UAH {
        name: string;
        symbol: string;
    }

    export interface TWD {
        name: string;
        symbol: string;
    }

    export interface CKD {
        name: string;
        symbol: string;
    }

    export interface ANG {
        name: string;
        symbol: string;
    }

    export interface ISK {
        name: string;
        symbol: string;
    }

    export interface PYG {
        name: string;
        symbol: string;
    }

    export interface SZL {
        name: string;
        symbol: string;
    }

    export interface HUF {
        name: string;
        symbol: string;
    }

    export interface MDL {
        name: string;
        symbol: string;
    }
// ...
    export interface Currencies {
        EUR: EUR;
        NZD: NZD;
        CUC: CUC;
        CUP: CUP;
        VES: VES;
        USD: USD;
        WST: WST;
        KHR: KHR;
        HKD: HKD;
        MRU: MRU;
        YER: YER;
        DJF: DJF;
        EGP: EGP;
        HRK: HRK;
        CHF: CHF;
        KZT: KZT;
        DKK: DKK;
        XOF: XOF;
        BMD: BMD;
        SHP: SHP;
        ZAR: ZAR;
        RWF: RWF;
        MOP: MOP;
        BIF: BIF;
        UAH: UAH;
        TWD: TWD;
        CKD: CKD;
        ANG: ANG;
        ISK: ISK;
        PYG: PYG;
        SZL: SZL;
        HUF: HUF;
        MDL: MDL;
        CLP: CLP;
        AUD: AUD;
        UYU: UYU;
        LKR: LKR;
        GTQ: GTQ;
        GHS: GHS;
        ILS: ILS;
        MZN: MZN;
        BTN: BTN;
        INR: INR;
        KYD: KYD;
        KPW: KPW;
        BHD: BHD;
        FOK: FOK;
        AWG: AWG;
        IQD: IQD;
        MAD: MAD;
        MKD: MKD;
        PLN: PLN;
        SBD: SBD;
        BRL: BRL;
        OMR: OMR;
        THB: THB;
        XAF: XAF;
        AMD: AMD;
        HNL: HNL;
        ZMW: ZMW;
        NOK: NOK;
        COP: COP;
        BBD: BBD;
        LYD: LYD;
        RSD: RSD;
        SEK: SEK;
        AOA: AOA;
        PAB: PAB;
        MUR: MUR;
        TZS: TZS;
        GEL: GEL;
        GMD: GMD;
        ZWL: ZWL;
        BZD: BZD;
        XCD: XCD;
        BYN: BYN;
        CAD: CAD;
        ARS: ARS;
        SRD: SRD;
        GNF: GNF;
        SSP: SSP;
        NAD: NAD;
        QAR: QAR;
        MMK: MMK;
        FKP: FKP;
        STN: STN;
        CDF: CDF;
        PHP: PHP;
        BND: BND;
        SGD: SGD;
        SLL: SLL;
        MNT: MNT;
        DZD: DZD;
        SDG: SDG;
        ETB: ETB;
        BAM: BAM;
        KMF: KMF;
        KWD: KWD;
        XPF: XPF;
        BDT: BDT;
        JMD: JMD;
        GIP: GIP;
        MYR: MYR;
        GBP: GBP;
        MGA: MGA;
        TVD: TVD;
        RUB: RUB;
        VND: VND;
        NPR: NPR;
        KID: KID;
        LRD: LRD;
        SOS: SOS;
        RON: RON;
        TND: TND;
        AZN: AZN;
        PGK: PGK;
        LSL: LSL;
        BSD: BSD;
        CNY: CNY;
        AED: AED;
        TMT: TMT;
        LAK: LAK;
        JOD: JOD;
        SCR: SCR;
        UZS: UZS;
        NIO: NIO;
        KES: KES;
        JEP: JEP;
        CZK: CZK;
        UGX: UGX;
        CRC: CRC;
        MXN: MXN;
        ERN: ERN;
        JPY: JPY;
        VUV: VUV;
        BWP: BWP;
        TOP: TOP;
        FJD: FJD;
        ALL: ALL;
        PEN: PEN;
        IDR: IDR;
        GGP: GGP;
        KGS: KGS;
        BOB: BOB;
        DOP: DOP;
        LBP: LBP;
        MVR: MVR;
        KRW: KRW;
        SYP: SYP;
        AFN: AFN;
        MWK: MWK;
        TJS: TJS;
        TRY: TRY;
        NGN: NGN;
        PKR: PKR;
        IRR: IRR;
        CVE: CVE;
        BGN: BGN;
        HTG: HTG;
        IMP: IMP;
        GYD: GYD;
        SAR: SAR;
        TTD: TTD;
    }

    export interface Idd {
        root: string;
        suffixes: string[];
    }

    export interface Languages {
        cnr: string;
        eng: string;
        smo: string;
        tkl: string;
        spa: string;
        fra: string;
        ell: string;
        khm: string;
        zho: string;
        ara: string;
        hrv: string;
        deu: string;
        kaz: string;
        rus: string;
        dan: string;
        cal: string;
        cha: string;
        ita: string;
        afr: string;
        nbl: string;
        nso: string;
        sot: string;
        ssw: string;
        tsn: string;
        tso: string;
        ven: string;
        xho: string;
        zul: string;
        kin: string;
        por: string;
        run: string;
        ukr: string;
        rar: string;
        pov: string;
        nld: string;
        isl: string;
        grn: string;
        hun: string;
        ron: string;
        kal: string;
        nau: string;
        sin: string;
        tam: string;
        heb: string;
        dzo: string;
        kor: string;
        fao: string;
        pap: string;
        arc: string;
        ckb: string;
        ber: string;
        mkd: string;
        pol: string;
        slv: string;
        tha: string;
        sag: string;
        hye: string;
        nor: string;
        ltz: string;
        srp: string;
        swe: string;
        mfe: string;
        swa: string;
        kat: string;
        mlt: string;
        bwg: string;
        kck: string;
        khi: string;
        ndc: string;
        nde: string;
        nya: string;
        sna: string;
        toi: string;
        zib: string;
        bjz: string;
        lat: string;
        bel: string;
        her: string;
        hgm: string;
        kwn: string;
        loz: string;
        ndo: string;
        mya: string;
        gle: string;
        lit: string;
        kon: string;
        lin: string;
        lua: string;
        fil: string;
        msa: string;
        mon: string;
        mey: string;
        tet: string;
        fin: string;
        bar: string;
        amh: string;
        bos: string;
        zdj: string;
        nno: string;
        nob: string;
        smi: string;
        ben: string;
        jam: string;
        cat: string;
        mlg: string;
        gsw: string;
        roh: string;
        tvl: string;
        vie: string;
        nep: string;
        gil: string;
        som: string;
        mah: string;
        aze: string;
        hmo: string;
        tpi: string;
        tur: string;
        tuk: string;
        lao: string;
        crs: string;
        uzb: string;
        nrf: string;
        ces: string;
        slk: string;
        mri: string;
        nzs: string;
        tir: string;
        jpn: string;
        sqi: string;
        bis: string;
        pau: string;
        ton: string;
        fij: string;
        hif: string;
        niu: string;
        aym: string;
        que: string;
        ind: string;
        nfr: string;
        kir: string;
        div: string;
        pih: string;
        prs: string;
        pus: string;
        tgk: string;
        hin: string;
        urd: string;
        lav: string;
        fas: string;
        bul: string;
        est: string;
        hat: string;
        glv: string;
    }
// ...
    export interface Translations {
        ara: Ara2;
        ces: Ces2;
        cym: Cym;
        deu: Deu2;
        est: Est2;
        fin: Fin2;
        fra: Fra2;
        hrv: Hrv2;
        hun: Hun2;
        ita: Ita2;
        jpn: Jpn2;
        kor: Kor2;
        nld: Nld2;
        per: Per;
        pol: Pol2;
        por: Por2;
        rus: Rus2;
        slk: Slk2;
        spa: Spa2;
        swe: Swe2;
        urd: Urd2;
        zho: Zho2;
    }

    export interface Eng2 {
        f: string;
        m: string;
    }

    export interface Fra3 {
        f: string;
        m: string;
    }

    export interface Demonyms {
        eng: Eng2;
        fra: Fra3;
    }

    export interface Maps {
        googleMaps: string;
        openStreetMaps: string;
    }

    export interface Gini {
        2016: number;
        2018?: number;
        2006?: number;
        2013?: number;
        2014?: number;
        2017?: number;
        2015?: number;
        2019?: number;
        2010?: number;
        2012?: number;
        2008?: number;
        2009?: number;
        1999?: number;
        2011?: number;
        2004?: number;
        2005?: number;
        1998?: number;
        2003?: number;
        1992?: number;
    }

    export interface Car {
        signs: string[];
        side: string;
    }

    export interface Flags {
        png: string;
        svg: string;
    }

    export interface CoatOfArms {
        png: string;
        svg: string;
    }

    export interface CapitalInfo {
        latlng: number[];
    }

    export interface PostalCode {
        format: string;
        regex: string;
    }

    export interface CountryObject {
        name: Name;
        tld: string[];
        cca2: string;
        ccn3: string;
        cca3: string;
        cioc: string;
        independent: boolean;
        status: string;
        unMember: boolean;
        currencies: Currencies;
        idd: Idd;
        capital: string[];
        altSpellings: string[];
        region: string;
        subregion: string;
        languages: Languages;
        translations: Translations;
        latlng: number[];
        landlocked: boolean;
        borders: string[];
        area: number;
        demonyms: Demonyms;
        flag: string;
        maps: Maps;
        population: number;
        gini: Gini;
        fifa: string;
        car: Car;
        timezones: string[];
        continents: string[];
        flags: Flags;
        coatOfArms: CoatOfArms;
        startOfWeek: string;
        capitalInfo: CapitalInfo;
        postalCode: PostalCode;
    }